@alanorwick 2019-12-02
Hi @mattjmcnaughton! Thank you, very excited to have the chance to develop on Kubernetes. It's been fun. Just updated the body of the PR to reflect more info on the expected behavior. Looking in to unit tests as well.