@mattjmcnaughton 2019-12-04
That makes sense to me assuming that we do need a jitter. But curious to hear
from some of the others tagged here :)