@ryanarifin134 2019-12-13
That sounds great! I'll take care of it. For the jitter range, I was thinking we could base it off a percentage of the `InitialDelaySeconds` value. How does 25% of the `InitialDelaySeconds` value sound?