@mattjmcnaughton 2019-12-14
> That sounds great! I'll take care of it. For the jitter range, I was thinking we could base it off a percentage of the `InitialDelaySeconds` value. How does 25% of the `InitialDelaySeconds` value sound?

That sounds good to me :) Let's design it so its easy to refine in the future should we want a different value.