@tedyu 2020-05-13
This is useful fix which makes probe behavior less surprising.
Can this be reopened ?